requirements.txt
# Core Deep Learning
torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0

# Computer Vision
opencv-python-headless==4.12.0.88
pillow==11.3.0

numpy==2.2.6

# Video Processing
decord==0.6.0

# Model Architecture
timm==1.0.21

# Data Augmentation
albumentations==2.0.8
albucore==0.0.24

# Face Detection
mediapipe>=0.10.0

# Visualization
matplotlib==3.10.6
imageio==2.34.0
imageio-ffmpeg>=0.4.8

# Machine Learning Utilities
scikit-learn>=1.4.0

# Progress Bars
tqdm==4.67.1

# Experiment Tracking (Optional)
wandb==0.16.0

# Configuration
PyYAML==6.0.3
